+++ src/clientfactory/core/utils/request/__init__.py
@@ -1,0 +1,8 @@
+# ~/clientfactory/src/clientfactory/core/utils/request/__init__.py
+from .path import (
+    resolveargs, substitute
+)
+from .building import (
+    separatekwargs, buildrequest,
+    applymethodconfig
+)